fix bugs + tree view
